split resolvers
