+++ src/resolver/single.js
@@ -1,0 +1,1 @@
+export default entityData => ({ id }) => entityData.find(d => d.id == id);
