+++ src/resolver/update.js
@@ -1,0 +1,10 @@
+export default entityData => ({ id, ...updates }) => {
+    const parsedId = parseInt(id, 10); // FIXME fails for non-integer ids
+    const indexOfEntity = entityData.findIndex(e => e.id === parsedId);
+
+    entityData[indexOfEntity] = {
+        ...entityData[indexOfEntity],
+        ...updates,
+    };
+    return entityData[indexOfEntity];
+};
